Create MatchTheShoe.py
The popularity order of an item is defined as follows:

Item  is more popular than item  if it was ordered more times than item .
If item  and item  were ordered an equal number of times, then the item with a smaller ID (see below) is more popular.
Choosing the right pair of shoes to purchase with an article of clothing is not an obvious decision. Given the purchase frequency for shoes that were previously purchased with an item, can you help a customer by showing them the  most popular shoes (in popularity order)?

Input Format

The first line contains three space-separated integers describing the following respective values:

 - The number of most popular shoes you must suggest.
 - The number of distinct shoe IDs.
 - The number of orders.
Each line  of the  subsequent lines (where ) contains a single integer, , denoting the ID number of shoes that were previously ordered with the item.

Constraints





Output Format

Sort your list of  shoe suggestions in order of highest to lowest popularity, and then print each element of the prioritized list on a new line.

Sample Input

3 4 8
2
1
2
0
3
3
1
2
Sample Output

2
1
3
Explanation

The shoes with  appeared  times. Because this is the most popular item, we print it first.
The shoes with  and  appear  times each. Because both shoes appear the same number of times, they are prioritized by ascending  number and printed on the second and third lines.
Shoe  is not part of this list, as its popularity rank is less than .