Create TheInquiringManager.py
You are responsible for auditing all customer orders. Each order has a price, , and a timestamp, , denoting the minute when it was received.

At any given time, your manager can ask you for the maximum price of any order placed within the last minutes. For example, if your manager stops in at minute , you must know the maximum price of any order received between minutes  and . Observe that you must also account for orders received in the same minute as your manager arrived.

You are given a list of  events, where each event represents either a received order or an inquiry. For each inquiry from your manager, print the maximum price for any order placed within the last  minutes on a new line.

Input Format

The first line contains a single integer, , denoting the number of events. Each of the  subsequent lines describes either of the following types of events:

1 P T - An order with price  was placed at time .
2 T - Your manager stopped in at time  to inquire about the maximum order price for the last  minutes.
All the events are given in non-decreasing order with respect to , so for , we have .

If there are multiple events of different types occurring at the same time, the received orders will be listed before manager inquiries.


Output Format

For each event of type , print a number denoting the maximum price for any order placed within the previous  minutes on a new line. If no orders were placed during that interval, print .

Assume there will be at least one event of type .

Sample Input

11
1 150 0
1 3 10
2 40
1 143 59
2 59
1 100 60
2 60
1 159 61
2 61
2 120
2 121
Sample Output

150
150
143
159
159
-1
Explanation

The largest order price in time range  is .
The largest order price in time range  is still .
The largest order price in time range  is .
The largest order price in time range  is .
The largest order price in time range  is still .
There were no orders placed in time range , so we print  for this inquiry.