Create MinimalWrappingSurfaceArea.py
When an item is out of stock in one warehouse, it's typically replenished through transfer from another warehouse. In this challenge, we focus on stock items enclosed in boxes of the same size. To ensure safe shipment, we want to align all boxes in a shipment so that the surface area of a bounding box is minimal.

Given a set of  boxes, all of which have identical  dimensions, print the minimum surface area, , for a bounding box to transport them. The boxes can be stacked arbitrarily, but their rotation along all axes must be the same.